Refactor validate method to EntitySubscription
