+++ snuba/subscriptions/entity_subscription.py
@@ -1,12 +1,20 @@
 from abc import ABC, abstractmethod
 from enum import Enum
-from typing import Any, List, Mapping, Optional, Sequence, Type
+from typing import Any, Mapping, Optional, Sequence, Type, Union
 
 from snuba.datasets.entities import EntityKey
+from snuba.datasets.entities.factory import get_entity
+from snuba.query.composite import CompositeQuery
 from snuba.query.conditions import ConditionFunctions, binary_condition
+from snuba.query.data_source.simple import Entity
 from snuba.query.exceptions import InvalidQueryException
 from snuba.query.expressions import Column, Expression, FunctionCall, Literal
+from snuba.query.logical import Query
 from snuba.query.types import Condition
+from snuba.query.validation.validators import (
+    NoTimeBasedConditionValidator,
+    SubscriptionAllowedClausesValidator,
+)
 
 
 class SubscriptionType(Enum):
@@ -15,16 +23,18 @@ class SubscriptionType(Enum):
     DELEGATE = "delegate"
 
 
-class EntitySubscription(ABC):
-    MAX_ALLOWED_AGGREGATIONS: int = 1
+class InvalidSubscriptionError(Exception):
+    pass
 
+
+class EntitySubscription(ABC):
     def __init__(self, data_dict: Mapping[str, Any]) -> None:
         ...
 
     @abstractmethod
     def get_entity_subscription_conditions_for_snql(
         self, offset: Optional[int] = None
-    ) -> List[Expression]:
+    ) -> Sequence[Expression]:
         """
         Returns a list of extra conditions that are entity specific and required for the
         snql subscriptions
@@ -42,11 +52,35 @@ class EntitySubscription(ABC):
         raise NotImplementedError
 
     @abstractmethod
+    def validate_query(self, query: Union[CompositeQuery[Entity], Query]) -> None:
+        """
+        Applies entity specific validations on query argument passed
+        """
+        raise NotImplementedError
+
+    @abstractmethod
     def to_dict(self) -> Mapping[str, Any]:
         raise NotImplementedError
 
 
-class SessionsSubscription(EntitySubscription):
+class EntitySubscriptionValidation:
+    MAX_ALLOWED_AGGREGATIONS: int = 1
+
+    def validate_query(self, query: Union[CompositeQuery[Entity], Query]) -> None:
+        # TODO: Support composite queries with multiple entities.
+        from_clause = query.get_from_clause()
+        if not isinstance(from_clause, Entity):
+            raise InvalidSubscriptionError("Only simple queries are supported")
+        entity = get_entity(from_clause.key)
+
+        SubscriptionAllowedClausesValidator(self.MAX_ALLOWED_AGGREGATIONS).validate(
+            query
+        )
+        if entity.required_time_column:
+            NoTimeBasedConditionValidator(entity.required_time_column).validate(query)
+
+
+class SessionsSubscription(EntitySubscriptionValidation, EntitySubscription):
     MAX_ALLOWED_AGGREGATIONS: int = 2
 
     def __init__(self, data_dict: Mapping[str, Any]) -> None:
@@ -60,7 +94,7 @@ class SessionsSubscription(EntitySubscription):
 
     def get_entity_subscription_conditions_for_snql(
         self, offset: Optional[int] = None
-    ) -> List[Expression]:
+    ) -> Sequence[Expression]:
         return [
             binary_condition(
                 ConditionFunctions.EQ,
@@ -78,10 +112,10 @@ class SessionsSubscription(EntitySubscription):
         return {"organization": self.organization}
 
 
-class BaseEventsSubscription(EntitySubscription, ABC):
+class BaseEventsSubscription(EntitySubscriptionValidation, EntitySubscription, ABC):
     def get_entity_subscription_conditions_for_snql(
         self, offset: Optional[int] = None
-    ) -> List[Expression]:
+    ) -> Sequence[Expression]:
         if offset is None:
             return []
